Stop vendoring compiler tools for DEVELOPMENT SDKs
DEVELOPMENT SDKs were just copying compiler tools from the swift.org
toolchains. This was causing the SDKs to be much larger even though
the compiler tools are already available in the host installed toolchain.

This change makes the DEVELOPMENT SDKs not copy the compiler tools from
the base toolchain. Note that We still need to copy some of LLVM tools
that are not part of the base toolchain.