Sync date part of SDK version with the base tag
We were using the date SDK was built but it was not the same as the
date snapshot toolchain was built. This was usually causing the SDK
version to be off by one day from the toolchain version. Although SDK
version and toolchain version are not directly related and there is no
guarantee that they have 1:1 mapping (multiple SDK snapshots can be
built with the same toolchain snapshot), it is better to have them in
sync to avoid confusion.